+++ previous-versions/sync-for-expenses-version-1/expenses.go
@@ -14,20 +14,20 @@ import (
 	"net/http"
 )
 
-// expenses - Create expense datasets and upload receipts.
-type expenses struct {
+// Expenses - Create expense datasets and upload receipts.
+type Expenses struct {
 	sdkConfiguration sdkConfiguration
 }
 
-func newExpenses(sdkConfig sdkConfiguration) *expenses {
-	return &expenses{
+func newExpenses(sdkConfig sdkConfiguration) *Expenses {
+	return &Expenses{
 		sdkConfiguration: sdkConfig,
 	}
 }
 
 // CreateExpenseDataset - Create expense-transactions
 // Create an expense transaction
-func (s *expenses) CreateExpenseDataset(ctx context.Context, request operations.CreateExpenseDatasetRequest, opts ...operations.Option) (*operations.CreateExpenseDatasetResponse, error) {
+func (s *Expenses) CreateExpenseDataset(ctx context.Context, request operations.CreateExpenseDatasetRequest, opts ...operations.Option) (*operations.CreateExpenseDatasetResponse, error) {
 	o := operations.Options{}
 	supportedOptions := []string{
 		operations.SupportedOptionRetries,
@@ -140,15 +140,18 @@ func (s *expenses) CreateExpenseDataset(ctx context.Context, request operations.
 	case httpRes.StatusCode == 503:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out shared.ErrorMessage
+			var out sdkerrors.ErrorMessage
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
-
-			res.ErrorMessage = &out
+			return nil, &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
+	case httpRes.StatusCode >= 400 && httpRes.StatusCode < 500:
+		fallthrough
+	case httpRes.StatusCode >= 500 && httpRes.StatusCode < 600:
+		return nil, sdkerrors.NewSDKError("API error occurred", httpRes.StatusCode, string(rawBody), httpRes)
 	}
 
 	return res, nil
@@ -156,7 +159,7 @@ func (s *expenses) CreateExpenseDataset(ctx context.Context, request operations.
 
 // UpdateExpenseDataset - Update expense transactions
 // Update an expense transaction
-func (s *expenses) UpdateExpenseDataset(ctx context.Context, request operations.UpdateExpenseDatasetRequest, opts ...operations.Option) (*operations.UpdateExpenseDatasetResponse, error) {
+func (s *Expenses) UpdateExpenseDataset(ctx context.Context, request operations.UpdateExpenseDatasetRequest, opts ...operations.Option) (*operations.UpdateExpenseDatasetResponse, error) {
 	o := operations.Options{}
 	supportedOptions := []string{
 		operations.SupportedOptionRetries,
@@ -243,12 +246,12 @@ func (s *expenses) UpdateExpenseDataset(ctx context.Context, request operations.
 	case httpRes.StatusCode == 202:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out operations.UpdateExpenseDataset202ApplicationJSON
+			var out operations.UpdateExpenseDatasetResponseBody
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
 
-			res.UpdateExpenseDataset202ApplicationJSONObject = &out
+			res.Object = &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
@@ -271,15 +274,18 @@ func (s *expenses) UpdateExpenseDataset(ctx context.Context, request operations.
 	case httpRes.StatusCode == 503:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out shared.ErrorMessage
+			var out sdkerrors.ErrorMessage
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
-
-			res.ErrorMessage = &out
+			return nil, &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
+	case httpRes.StatusCode >= 400 && httpRes.StatusCode < 500:
+		fallthrough
+	case httpRes.StatusCode >= 500 && httpRes.StatusCode < 600:
+		return nil, sdkerrors.NewSDKError("API error occurred", httpRes.StatusCode, string(rawBody), httpRes)
 	}
 
 	return res, nil
@@ -287,7 +293,7 @@ func (s *expenses) UpdateExpenseDataset(ctx context.Context, request operations.
 
 // UploadAttachment - Upload attachment
 // Creates an attachment in the accounting software against the given transactionId
-func (s *expenses) UploadAttachment(ctx context.Context, request operations.UploadAttachmentRequest, opts ...operations.Option) (*operations.UploadAttachmentResponse, error) {
+func (s *Expenses) UploadAttachment(ctx context.Context, request operations.UploadAttachmentRequest, opts ...operations.Option) (*operations.UploadAttachmentResponse, error) {
 	o := operations.Options{}
 	supportedOptions := []string{
 		operations.SupportedOptionRetries,
@@ -400,15 +406,18 @@ func (s *expenses) UploadAttachment(ctx context.Context, request operations.Uplo
 	case httpRes.StatusCode == 503:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out shared.ErrorMessage
+			var out sdkerrors.ErrorMessage
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
-
-			res.ErrorMessage = &out
+			return nil, &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
+	case httpRes.StatusCode >= 400 && httpRes.StatusCode < 500:
+		fallthrough
+	case httpRes.StatusCode >= 500 && httpRes.StatusCode < 600:
+		return nil, sdkerrors.NewSDKError("API error occurred", httpRes.StatusCode, string(rawBody), httpRes)
 	}
 
 	return res, nil