+++ previous-versions/sync-for-expenses-version-1/pkg/models/operations/getsynctransaction.go
@@ -40,14 +40,12 @@ func (o *GetSyncTransactionRequest) GetTransactionID() string {
 type GetSyncTransactionResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
 	// Success
-	TransactionMetadata []shared.TransactionMetadata
+	Classes []shared.TransactionMetadata
 }
 
 func (o *GetSyncTransactionResponse) GetContentType() string {
@@ -55,13 +53,6 @@ func (o *GetSyncTransactionResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GetSyncTransactionResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GetSyncTransactionResponse) GetStatusCode() int {
@@ -78,9 +69,9 @@ func (o *GetSyncTransactionResponse) GetRawResponse() *http.Response {
 	return o.RawResponse
 }
 
-func (o *GetSyncTransactionResponse) GetTransactionMetadata() []shared.TransactionMetadata {
+func (o *GetSyncTransactionResponse) GetClasses() []shared.TransactionMetadata {
 	if o == nil {
 		return nil
 	}
-	return o.TransactionMetadata
+	return o.Classes
 }