+++ previous-versions/sync-for-expenses-version-1/transactionstatus.go
@@ -14,20 +14,20 @@ import (
 	"net/http"
 )
 
-// transactionStatus - Retrieve the status of transactions within a sync.
-type transactionStatus struct {
+// TransactionStatus - Retrieve the status of transactions within a sync.
+type TransactionStatus struct {
 	sdkConfiguration sdkConfiguration
 }
 
-func newTransactionStatus(sdkConfig sdkConfiguration) *transactionStatus {
-	return &transactionStatus{
+func newTransactionStatus(sdkConfig sdkConfiguration) *TransactionStatus {
+	return &TransactionStatus{
 		sdkConfiguration: sdkConfig,
 	}
 }
 
 // GetSyncTransaction - Get sync transaction
 // Gets the status of a transaction for a sync
-func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request operations.GetSyncTransactionRequest, opts ...operations.Option) (*operations.GetSyncTransactionResponse, error) {
+func (s *TransactionStatus) GetSyncTransaction(ctx context.Context, request operations.GetSyncTransactionRequest, opts ...operations.Option) (*operations.GetSyncTransactionResponse, error) {
 	o := operations.Options{}
 	supportedOptions := []string{
 		operations.SupportedOptionRetries,
@@ -112,7 +112,7 @@ func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request oper
 				return nil, err
 			}
 
-			res.TransactionMetadata = out
+			res.Classes = out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
@@ -131,15 +131,18 @@ func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request oper
 	case httpRes.StatusCode == 503:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out shared.ErrorMessage
+			var out sdkerrors.ErrorMessage
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
-
-			res.ErrorMessage = &out
+			return nil, &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
+	case httpRes.StatusCode >= 400 && httpRes.StatusCode < 500:
+		fallthrough
+	case httpRes.StatusCode >= 500 && httpRes.StatusCode < 600:
+		return nil, sdkerrors.NewSDKError("API error occurred", httpRes.StatusCode, string(rawBody), httpRes)
 	}
 
 	return res, nil
@@ -147,7 +150,7 @@ func (s *transactionStatus) GetSyncTransaction(ctx context.Context, request oper
 
 // ListSyncTransactions - Get sync transactions
 // Get's the transactions and status for a sync
-func (s *transactionStatus) ListSyncTransactions(ctx context.Context, request operations.ListSyncTransactionsRequest, opts ...operations.Option) (*operations.ListSyncTransactionsResponse, error) {
+func (s *TransactionStatus) ListSyncTransactions(ctx context.Context, request operations.ListSyncTransactionsRequest, opts ...operations.Option) (*operations.ListSyncTransactionsResponse, error) {
 	o := operations.Options{}
 	supportedOptions := []string{
 		operations.SupportedOptionRetries,
@@ -257,15 +260,18 @@ func (s *transactionStatus) ListSyncTransactions(ctx context.Context, request op
 	case httpRes.StatusCode == 503:
 		switch {
 		case utils.MatchContentType(contentType, `application/json`):
-			var out shared.ErrorMessage
+			var out sdkerrors.ErrorMessage
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
-
-			res.ErrorMessage = &out
+			return nil, &out
 		default:
 			return nil, sdkerrors.NewSDKError(fmt.Sprintf("unknown content-type received: %s", contentType), httpRes.StatusCode, string(rawBody), httpRes)
 		}
+	case httpRes.StatusCode >= 400 && httpRes.StatusCode < 500:
+		fallthrough
+	case httpRes.StatusCode >= 500 && httpRes.StatusCode < 600:
+		return nil, sdkerrors.NewSDKError("API error occurred", httpRes.StatusCode, string(rawBody), httpRes)
 	}
 
 	return res, nil